Fedora暫定対応。setenforceコマンドの順番変更

+++ centos6-rdo.sh
@@ -1,15 +1,27 @@
 #!/bin/bash -x
 
-#RDO Install Assist v.140525-1845
+#RDO Install Assist v.140526-1200
 #
-yum install vim tracerourte system-config-firewall-tui system-config-network-tui
+# ディストリビューション名とバージョンを取得する(参考サイト)
+#http://geektrainee.hatenablog.jp/entry/2013/11/27/022633
 
-read -p "Do you want to Copy the sysctl.conf(y/n)?"
-[ "$REPLY" == "y" ] && (cp conf/sysctl.conf /etc/sysctl.conf;sysctl -e -p /etc/sysctl.conf)
-[ "$REPLY" == "n" ] && Skipped!
+tmp=`cat /etc/issue | head -n 1`
+DIST=`echo $tmp | awk '{print $1}'`
+
+if [[ $DIST =~ "CentOS" ]]; then
+  yum install vim tracerourte system-config-firewall-tui system-config-network-tui
+  read -p "Do you want to Copy the sysctl.conf(y/n)?"
+  [ "$REPLY" == "y" ] && (cp conf/sysctl.conf /etc/sysctl.conf;sysctl -e -p /etc/sysctl.conf)
+  [ "$REPLY" == "n" ] && Skipped!
+elif [[ $DIST =~ "Fedora" ]]; then
+  yum install vim tracerourte system-config-network-tui
+  read -p "Do you want to Copy the sysctl.conf(y/n)?"
+  [ "$REPLY" == "y" ] && (cat conf/source.txt >> /etc/sysctl.conf;sysctl -e -p /etc/sysctl.conf)
+  [ "$REPLY" == "n" ] && Skipped!
+fi
 
 read -p "Do you want to Set SELinux(y/n)?"
-[ "$REPLY" == "y" ] && (sed -i -e s/^SELINUX=.*/SELINUX=permissive/ /etc/selinux/config;setenforce 0)
+[ "$REPLY" == "y" ] && (setenforce 0;sed -i -e s/^SELINUX=.*/SELINUX=permissive/ /etc/selinux/config)
 [ "$REPLY" == "n" ] && Skipped!
 
 read -p "Set the Repo(havana/icehouse/skip)?"
@@ -22,5 +34,5 @@ yum install -y openstack-packstack python-netaddr
 
 read -p "Do you want to Custom installation of RDO OpenStack(y/n)?"
 [ "$REPLY" == "y" ] && (packstack --gen-answer-file=answer.txt;echo "Edit the answer.txt File,After that Run the packstack command.")
-[ "$REPLY" == "n" ] && packstack --allinone
+[ "$REPLY" == "n" ] && (packstack --allinone;ln -s keystonerc_admin openrc)
